test(e2e): split daily flows and add handoff coverage
